♻️ refactor: Move webhook logic to app module
- Move webhook functionality to the app module
- Remove unnecessary webhook module, controller, and service files
- Create AppController to handle webhook requests
- Update AppService to include webhook logic
- Move CreateWebhookDto to the root dto folder
