--- src/webhook/entities/webhook.entity.ts
@@ -1,1 +1,0 @@
-export class Webhook {}
