--- src/webhook/webhook.controller.spec.ts
@@ -1,20 +1,0 @@
-import { Test, TestingModule } from '@nestjs/testing';
-import { WebhookController } from './webhook.controller';
-import { WebhookService } from './webhook.service';
-
-describe('WebhookController', () => {
-  let controller: WebhookController;
-
-  beforeEach(async () => {
-    const module: TestingModule = await Test.createTestingModule({
-      controllers: [WebhookController],
-      providers: [WebhookService],
-    }).compile();
-
-    controller = module.get<WebhookController>(WebhookController);
-  });
-
-  it('should be defined', () => {
-    expect(controller).toBeDefined();
-  });
-});
