--- src/webhook/webhook.controller.ts
@@ -1,13 +1,0 @@
-import { Controller, Post, Body } from '@nestjs/common';
-import { WebhookService } from './webhook.service';
-import { CreateWebhookDto } from './dto/create-webhook.dto';
-
-@Controller('webhook')
-export class WebhookController {
-  constructor(private readonly webhookService: WebhookService) {}
-
-  @Post()
-  create(@Body() createWebhookDto: CreateWebhookDto) {
-    return this.webhookService.create(createWebhookDto);
-  }
-}
