--- src/webhook/webhook.module.ts
@@ -1,11 +1,0 @@
-import { Module } from '@nestjs/common';
-import { WebhookService } from './webhook.service';
-import { WebhookController } from './webhook.controller';
-import { PrismaModule } from 'src/prisma/prisma.module';
-
-@Module({
-  imports: [PrismaModule],
-  controllers: [WebhookController],
-  providers: [WebhookService],
-})
-export class WebhookModule {}
